+++ account/transaction.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/Masterminds/semver/v3"
 	"github.com/NethermindEth/juno/core/felt"
 	"github.com/NethermindEth/starknet.go/client/rpcerr"
 	"github.com/NethermindEth/starknet.go/contracts"
@@ -59,8 +60,9 @@ func (account *Account) BuildAndSendInvokeTxn(
 		callData,
 		makeResourceBoundsMapWithZeroValues(),
 		&utils.TxnOptions{
-			Tip:         tip,
-			UseQueryBit: opts.UseQueryBit,
+			Tip:            tip,
+			UseQueryBit:    opts.UseQueryBit,
+			UseBlake2sHash: false,
 		},
 	)
 
@@ -138,6 +140,16 @@ func (account *Account) BuildAndSendDeclareTxn(
 		return response, err
 	}
 
+	var useBlake2sHash bool
+	if opts.UseBlake2sHash == nil {
+		useBlake2sHash, err = shouldUseBlake2sHash(ctx, account.Provider)
+		if err != nil {
+			return response, fmt.Errorf("failed to check whether to use Blake2s hash: %w", err)
+		}
+	} else {
+		useBlake2sHash = *opts.UseBlake2sHash
+	}
+
 	// building and signing the txn, as it needs a signature to estimate the fee
 	broadcastDeclareTxnV3, err := utils.BuildDeclareTxn(
 		account.Address,
@@ -146,8 +158,9 @@ func (account *Account) BuildAndSendDeclareTxn(
 		nonce,
 		makeResourceBoundsMapWithZeroValues(),
 		&utils.TxnOptions{
-			Tip:         tip,
-			UseQueryBit: opts.UseQueryBit,
+			Tip:            tip,
+			UseQueryBit:    opts.UseQueryBit,
+			UseBlake2sHash: useBlake2sHash,
 		},
 	)
 	if err != nil {
@@ -239,8 +252,9 @@ func (account *Account) BuildAndEstimateDeployAccountTxn(
 		classHash,
 		makeResourceBoundsMapWithZeroValues(),
 		&utils.TxnOptions{
-			Tip:         tip,
-			UseQueryBit: opts.UseQueryBit,
+			Tip:            tip,
+			UseQueryBit:    opts.UseQueryBit,
+			UseBlake2sHash: false,
 		},
 	)
 
@@ -464,4 +478,36 @@ func (account *Account) WaitForTransactionReceipt(
 			return receiptWithBlockInfo, nil
 		}
 	}
+}
+
+// shouldUseBlake2sHash determines whether to use the Blake2s hash function for the
+// compiled class hash.
+// Starknet v0.14.1 upgrade will deprecate the Poseidon hash function for the compiled class hash.
+//
+// Parameters:
+//   - ctx: The context
+//   - provider: The provider
+//
+// Returns:
+//   - bool: whether to use the Blake2s hash function for the compiled class hash
+//   - error: an error if any
+func shouldUseBlake2sHash(ctx context.Context, provider rpc.RPCProvider) (bool, error) {
+	block, err := provider.BlockWithTxHashes(ctx, rpc.WithBlockTag(rpc.BlockTagLatest))
+	if err != nil {
+		return false, fmt.Errorf("failed to get block with tx hashes: %w", err)
+	}
+
+	blockTxHashes, ok := block.(*rpc.BlockTxHashes)
+	if !ok {
+		return false, fmt.Errorf("block is not a BlockTxHashes: %T", block)
+	}
+
+	upgradeVersion := semver.MustParse("0.14.1")
+
+	currentVersion, err := semver.NewVersion(blockTxHashes.StarknetVersion)
+	if err != nil {
+		return false, fmt.Errorf("failed to parse block's starknet version: %w", err)
+	}
+
+	return currentVersion.Compare(upgradeVersion) >= 0, nil
 }